Modify update zshrc and init.vim; also add coc-settings.json.
Made some major changes to the init.vim file to accomodate the CoC
plugin. Additionally, removed some unecessary plugins and rearranged the
structure of the file a bit. The only change to the zshrc file was the
addition of the asdf plugin and the removal of the pyenv and nvm
plugins.